Relicense under the Unlicense
The reason is that it's not clear whether it's OK to download and use a single
bash script without also copying the license file. Since this is one of the
intended installation options, it's better to relicense the project under a
license that doesn't require any copyright notices to be included. According to
https://choosealicense.com/appendix and https://spdx.org/licenses, the only
option that is both OSI and FSF approved is the Unlicense, i.e. a public domain
dedication. Also, https://unlicense.org advises the inclusion of a CLA for
external contributors, but I believe it's covered by the GitHub Terms of Service
https://docs.github.com/en/site-policy/github-terms/github-terms-of-service#6-contributions-under-repository-license